squash and merge
