+++ ui/src/index.tsx
@@ -1,8 +1,8 @@
+import App from "app";
+import "index.css";
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./app";
-import "./index.css";
-import reportWebVitals from "./reportWebVitals";
+import reportWebVitals from "reportWebVitals";
 
 ReactDOM.render(
   <React.StrictMode>
